kms test refactoring
