+++ kms/composer.json
@@ -10,6 +10,7 @@
         ]
     },
     "require-dev": {
-        "phpunit/phpunit": "~4"
+        "phpunit/phpunit": "^5",
+        "google/cloud-tools": "^0.8.5"
     }
 }
