+++ monitoring/composer.json
@@ -23,7 +23,7 @@
         ]
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.8",
-        "google/cloud-tools": "^0.6.9"
+        "phpunit/phpunit": "^5",
+        "google/cloud-tools": "^0.8.5"
     }
 }
